Move some codes from Image Gallery to Dataset Card for reuse.
